requirements.txt
openai>=1.0.0
fal-client
python-dotenv
